requirements.txt
celery==5.2.1
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-timezone-field==4.2.1
djangorestframework~=3.12.4
PyJWT==2.3.0
redis==4.0.2


jwt~=1.3.1
requests~=2.26.0